+++ tools/src/com.oracle.truffle.tools.chromeinspector/src/com/oracle/truffle/tools/chromeinspector/InspectorExecutionContext.java
@@ -38,6 +38,7 @@ import java.util.logging.Level;
 import com.oracle.truffle.api.TruffleLogger;
 import com.oracle.truffle.api.debug.DebugException;
 import com.oracle.truffle.api.debug.DebugValue;
+import com.oracle.truffle.api.debug.DebuggerSession;
 import com.oracle.truffle.api.instrumentation.EventBinding;
 import com.oracle.truffle.api.instrumentation.SourceFilter;
 import com.oracle.truffle.api.instrumentation.TruffleInstrument;
@@ -185,15 +186,15 @@ public final class InspectorExecutionContext {
         }
     }
 
-    synchronized RemoteObjectsHandler getRemoteObjectsHandler() {
+    public synchronized RemoteObjectsHandler getRemoteObjectsHandler() {
         if (roh == null) {
-            roh = new RemoteObjectsHandler(err);
+            roh = new RemoteObjectsHandler(this);
         }
         return roh;
     }
 
     public RemoteObject createAndRegister(DebugValue value, boolean generatePreview) {
-        RemoteObject ro = new RemoteObject(value, generatePreview, getErr());
+        RemoteObject ro = new RemoteObject(value, generatePreview, this);
         if (ro.getId() != null) {
             getRemoteObjectsHandler().register(ro);
         }
@@ -290,6 +291,16 @@ public final class InspectorExecutionContext {
     }
 
     /**
+     * Returns the current debugger session if debugging is on.
+     *
+     * @return the current debugger session, or <code>null</code>.
+     */
+    public DebuggerSession getDebuggerSession() {
+        ScriptsHandler handler = this.sch;
+        return (handler != null) ? handler.getDebuggerSession() : null;
+    }
+
+    /**
      * For test purposes only. Do not call from production code.
      */
     public static void resetIDs() {
@@ -318,7 +329,7 @@ public final class InspectorExecutionContext {
         this.customObjectFormatterEnabled = enabled;
     }
 
-    boolean isCustomObjectFormatterEnabled() {
+    public boolean isCustomObjectFormatterEnabled() {
         return this.customObjectFormatterEnabled;
     }