+++ tools/src/com.oracle.truffle.tools.chromeinspector/src/com/oracle/truffle/tools/chromeinspector/types/RemoteObject.java
@@ -37,6 +37,7 @@ import com.oracle.truffle.api.debug.DebugScope;
 import com.oracle.truffle.api.debug.DebugValue;
 import com.oracle.truffle.api.instrumentation.TruffleInstrument;
 import com.oracle.truffle.api.nodes.LanguageInfo;
+import com.oracle.truffle.tools.chromeinspector.InspectorExecutionContext;
 import com.oracle.truffle.tools.chromeinspector.objects.NullObject;
 import com.oracle.truffle.tools.chromeinspector.types.TypeInfo.TYPE;
 
@@ -51,24 +52,25 @@ public final class RemoteObject {
     private final DebugScope valueScope;
     private final boolean generatePreview;
     private final String objectId;
-    private PrintWriter err = null;
+    private final InspectorExecutionContext context;
     private TypeInfo typeInfo;
     private Object value;
     private boolean replicableValue;
     private String unserializableValue;
     private String description;
     private JSONObject preview;
+    private JSONObject customPreview;
     private JSONObject jsonObject;
 
-    public RemoteObject(DebugValue debugValue, boolean generatePreview, PrintWriter err) {
-        this(debugValue, false, generatePreview, err);
+    public RemoteObject(DebugValue debugValue, boolean generatePreview, InspectorExecutionContext context) {
+        this(debugValue, false, generatePreview, context);
     }
 
-    public RemoteObject(DebugValue debugValue, boolean readEagerly, boolean generatePreview, PrintWriter err) {
+    public RemoteObject(DebugValue debugValue, boolean readEagerly, boolean generatePreview, InspectorExecutionContext context) {
         this.valueValue = debugValue;
         this.valueScope = null;
         this.generatePreview = generatePreview;
-        this.err = err;
+        this.context = context;
         if (!debugValue.hasReadSideEffects() || readEagerly) {
             boolean isObject = initFromValue();
             objectId = (isObject) ? Long.toString(LAST_ID.incrementAndGet()) : null;
@@ -85,6 +87,7 @@ public final class RemoteObject {
         if (originalLanguage != null) {
             debugValue = debugValue.asInLanguage(originalLanguage);
         }
+        PrintWriter err = context != null ? context.getErr() : null;
         this.typeInfo = TypeInfo.fromValue(debugValue, originalLanguage, err);
         String toString;
         Object rawValue = null;
@@ -136,6 +139,22 @@ public final class RemoteObject {
                 }
             }
         }
+        if (context != null && context.isCustomObjectFormatterEnabled()) {
+            if (originalLanguage != null) {
+                try {
+                    this.customPreview = CustomPreview.create(debugValue, originalLanguage, context);
+                } catch (DebugException ex) {
+                    if (err != null) {
+                        if (ex.isInternalError()) {
+                            err.println(debugValue.getName() + " custom preview has caused: " + ex);
+                            ex.printStackTrace(err);
+                        } else {
+                            err.println("Custom Formatter Failed: " + ex.getLocalizedMessage());
+                        }
+                    }
+                }
+            }
+        }
         return typeInfo.isObject;
     }
 
@@ -143,6 +162,7 @@ public final class RemoteObject {
         this.valueValue = null;
         this.valueScope = scope;
         this.generatePreview = false;
+        this.context = null;
         this.typeInfo = new TypeInfo(TYPE.OBJECT.getId(), null, null, null, true, false);
         this.value = null;
         this.replicableValue = false;
@@ -156,6 +176,7 @@ public final class RemoteObject {
         this.valueValue = null;
         this.valueScope = null;
         this.generatePreview = false;
+        this.context = null;
         this.typeInfo = new TypeInfo(type, subtype, className, null, true, false);
         this.value = null;
         this.replicableValue = false;
@@ -184,6 +205,7 @@ public final class RemoteObject {
         json.putOpt("description", description);
         json.putOpt("objectId", objectId);
         json.putOpt("preview", preview);
+        json.putOpt("customPreview", customPreview);
         return json;
     }